Refactor server architecture and add UI components
- Extracted connect endpoint logic to controllers/connect-controller.js
- Created token utilities module for JWT handling and PDA URL extraction
- Added web UI components (pages.js, styles.css) for success/error pages
- Updated Zoho API endpoints from .com to .eu domains
- Cleaned up server.js by removing inline business logic
- Removed obsolete test files and renamed dataswift to dataswyft
- Improved error handling and application token validation
