+++ .env.example
@@ -4,8 +4,8 @@
 # OAuth Configuration
 ZOHO_CRM_CLIENT_ID=your_client_id_here                    # Get from Zoho CRM Developer Console
 ZOHO_CRM_CLIENT_SECRET=your_client_secret_here            # Get from Zoho CRM Developer Console
-ZOHO_CRM_ACCOUNTS_URL=https://accounts.zoho.com
-ZOHO_CRM_API_DOMAIN=https://www.zohoapis.com
+ZOHO_CRM_ACCOUNTS_URL=https://accounts.zoho.eu
+ZOHO_CRM_API_DOMAIN=https://www.zohoapis.eu
 ZOHO_CRM_SCOPE=ZohoCRM.modules.contacts.ALL
 ZOHO_CRM_REDIRECT_URI=http://localhost:3000/oauth/callback
